- removed warning
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@762 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/D3D9Pad.cpp
@@ -448,7 +448,7 @@ void D3D9Pad::WrapOneLine (char* str, int len, int maxWidth)
 		     *it = pStr,  // 'iterator' char
 		     *pEnd = str + len, // <= point to terminating zero
 		     *pLastSpace = NULL;
-		int currentWidth = 0;
+		float currentWidth = 0;
 		while (it < pEnd)
 		{
 			while (it < pEnd && currentWidth < maxWidth) {
